chore (refs T22614): move RegisterFlyout to client/js.
