+++ demosplan/DemosPlanCoreBundle/Resources/client/js/bundles/publicDetail.js
@@ -17,7 +17,7 @@ import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
 import DpVideoPlayer from '@DpJs/components/core/DpVideoPlayer'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 import publicStatement from '@DemosPlanStatementBundle/store/PublicStatement'
-import RegisterFlyout from '@DemosPlanCoreBundle/components/RegisterFlyout'
+import RegisterFlyout from '@DpJs/components/core/RegisterFlyout'
 import StatementForm from '@DemosPlanStatementBundle/lib/StatementForm'
 import TableWrapper from '@DpJs/lib/TableWrapper'
