+++ demosplan/DemosPlanMiscContentBundle/Resources/client/js/bundles/dataProtection.js
@@ -8,7 +8,7 @@
  */
 
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import RegisterFlyout from '@DemosPlanCoreBundle/components/RegisterFlyout'
+import RegisterFlyout from '@DpJs/components/core/RegisterFlyout'
 
 const components = {
   RegisterFlyout
