+++ demosplan/DemosPlanUserBundle/Resources/client/js/bundles/alternativeLogin.js
@@ -15,7 +15,7 @@ import AlternativeLogin from '@DemosPlanUserBundle/components/AlternativeLogin'
 import DpDataTableExtended from '@DpJs/components/core/DpDataTable/DpDataTableExtended'
 import dpValidate from '@DpJs/lib/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import RegisterFlyout from '@DemosPlanCoreBundle/components/RegisterFlyout'
+import RegisterFlyout from '@DpJs/components/core/RegisterFlyout'
 import SamlLoginForm from '@DemosPlanUserBundle/components/samlLoginForm/SamlLoginForm'
 
 const components = {
