+++ demosplan/DemosPlanUserBundle/Resources/client/js/bundles/orgaRegisterForm.js
@@ -14,7 +14,7 @@
 import dpValidate from '@DpJs/lib/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 import OrgaRegisterForm from '@DemosPlanUserBundle/components/orgaRegisterForm/OrgaRegisterForm'
-import RegisterFlyout from '@DemosPlanCoreBundle/components/RegisterFlyout'
+import RegisterFlyout from '@DpJs/components/core/RegisterFlyout'
 
 const components = {
   OrgaRegisterForm,
