+++ demosplan/DemosPlanUserBundle/Resources/client/js/bundles/setPassword.js
@@ -13,7 +13,7 @@
 
 import dpValidate from '@DpJs/lib/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import RegisterFlyout from '@DemosPlanCoreBundle/components/RegisterFlyout'
+import RegisterFlyout from '@DpJs/components/core/RegisterFlyout'
 import SetPassword from '@DemosPlanUserBundle/components/portalUser/SetPassword'
 
 const components = {
